+++ cfgov/v1/preprocessed/js/routes/careers/application-process/index.js
@@ -1,7 +1,0 @@
-/* ==========================================================================
-   Scripts for `/careers/application-process/.
-   ========================================================================== */
-
-'use strict';
-
-require( '../../../modules/init-mobile-only-expandables' ).init();